fix(core): 修复潜在的整数溢出问题
在打印查询结果持续时间时，将 duration_us.count() 的结果显式转换为 int64_t 类型，以确保当持续时间超过 INT_MAX 时不会发生整数溢出。

+++ src/core/stream/stream_manager.cpp
@@ -391,7 +391,9 @@ void StreamManager::_on_query_result(const a_hashmap<uuids::uuid, ObjectData> &d
 
 	auto duration_us = std::chrono::duration_cast<std::chrono::microseconds>(end - start);
 
-	UtilityFunctions::print("[StreamManager] query_result: ", duration_us.count(), " us");
+	UtilityFunctions::print("[StreamManager] query_result: ", 
+                        static_cast<int64_t>(duration_us.count()), 
+                        " us");
 }
 
 void StreamManager::_load_object_scene(const uuids::uuid &uuid) {
